Group Actor tests in single file.
test_actor_lookup was not testing only the lookup but mainly
all the methods of the actor except the label which was tested
in a separate file.
